add confidence filter

+++ SN_CroppingTool.py
@@ -17,34 +17,37 @@ def normalize_2d(matrix):
 
 
 
-BaseSNR_Folder = input("Please enter images folder: Default is C:\Working\FindIMage_In_Dat\OutputTestSNR\ProcessSingles")
+BaseSNR_Folder = input("Please enter images folder: Default is C:\Working\FindIMage_In_Dat\Output")
 if len(BaseSNR_Folder)==0:
-    BaseSNR_Folder = r"C:\Working\FindIMage_In_Dat\OutputTestSNR\ProcessSingles"
+    BaseSNR_Folder = r"C:\Working\FindIMage_In_Dat\Output"
 
 OutputFolder=BaseSNR_Folder +"\\FilterOutput"
 print("Creating output folder",OutputFolder)
-_3DVisLabLib.DeleteFiles_RecreateFolder(OutputFolder)
+if _3DVisLabLib.yesno("Delete output folder?"):
+    _3DVisLabLib.DeleteFiles_RecreateFolder(OutputFolder)
 
 FolderFiles=_3DVisLabLib.GetAllFilesInFolder_Recursive(BaseSNR_Folder)
 #filter out non images
 ListAllImages=_3DVisLabLib.GetList_Of_ImagesInList(FolderFiles)
-
+print("ListAllImages",len(ListAllImages))
 FilteredToImage=[]
 for imgitem in ListAllImages:
-    if  not "d6" in imgitem.lower():continue
+    if  not "d8" in imgitem.lower():continue
     FilteredToImage.append(imgitem)
-
+print("FilteredToImage",len(FilteredToImage))
 #roll through all images and build up stacked image
 BaseImage_np=None
 kernel = np.ones((5,5),np.float32)/25#kernel size for smoothing - maybe make smoother as such small images
 for Indexer,Image in enumerate(FilteredToImage):
-    
+    #try:
     #load image
     AddImage=cv2.imread(Image,cv2.IMREAD_GRAYSCALE)
     #invert
     #AddImage = np.invert(AddImage)
     #blur
-    AddImage = cv2.filter2D(AddImage,-1,kernel)
+    #AddImage = cv2.filter2D(AddImage,-1,kernel)
+    #dilate
+    #AddImage = cv2.dilate(AddImage, kernel, iterations=1)
     #convert to numpy array to stop opencv limiting ceiling
     AddImage_np=np.asarray(AddImage,dtype="int32")
     #try clipping maximum grayscale value so white doesnt saturate imag
@@ -58,8 +61,9 @@ for Indexer,Image in enumerate(FilteredToImage):
     #add images together
     if Indexer%1==0:#reduce image set to try and mitigate over blurring
         BaseImage_np=BaseImage_np+AddImage_np
-        print((BaseImage_np.max()))
-
+        #print((BaseImage_np.max()))
+    #except:
+        #pass
 #normalise to image grayscale range, as max range at moment will not be compatible with opencv or wont
 #be interpreted correctly
 BaseImage=normalize_2d(BaseImage_np)
@@ -71,6 +75,7 @@ FromBase=0
 FromLeft=0
 FromRight=0
 ImageToDisplay=BaseImage.copy()
+SaveOutDictionary=dict()
 while True:
     #general shape for reference ImageToDisplay[175,335] 0 is height 335 is width
     #ImageToDisplay=BaseImage.copy()
@@ -89,8 +94,34 @@ while True:
         FromLeft=FromLeft+1
     if User_keypress=="d":
         FromRight=FromRight+1
+
+    if User_keypress=="i":
+        FromTop=FromTop-1
+    if User_keypress=="k":
+        FromBase=FromBase-1
+    if User_keypress=="j":
+        FromLeft=FromLeft-1
+    if User_keypress=="l":
+        FromRight=FromRight-1
+
+
+
     if User_keypress=="t":
         #roll through all images that have been averaged together for a quick preview
-
+        SaveOutDictionary=dict()#clean this out
+        for Indexer2,Image2 in enumerate(FilteredToImage):
+            Display=cv2.imread(Image2,cv2.IMREAD_GRAYSCALE)
+            ImageToDisplay=Display.copy()#
+            ImageToDisplay[:,:]=255
+            ImageToDisplay[FromTop:(BaseImage.shape[0]-FromBase),FromLeft:(BaseImage.shape[1]-FromRight)]=Display[FromTop:(BaseImage.shape[0]-FromBase),FromLeft:(BaseImage.shape[1]-FromRight)]
+            _3DVisLabLib.ImageViewer_Quickv2_UserControl(ImageToDisplay,0,False,False)
+            SaveOutDictionary[Image2]=ImageToDisplay
     
     
+    if User_keypress=="x":
+        for imgitem in SaveOutDictionary:
+            
+            FilenameSansPath=imgitem.split("\\")[-1]
+            print("saving out",OutputFolder + "\\" + FilenameSansPath)
+            cv2.imwrite(OutputFolder + "\\" + FilenameSansPath,SaveOutDictionary[imgitem])
+        print("Finished saving")
